+++ cfgov/unprocessed/js/modules/ClearableInput.js
@@ -1,10 +1,7 @@
 'use strict';
 
-// Required polyfills for IE9.
-if ( !Modernizr.classlist ) { require( '../modules/polyfill/class-list' ); } // eslint-disable-line no-undef, global-require, no-inline-comments, max-len
-
 // Required modules.
-var atomicCheckers = require( '../modules/util/atomic-checkers' );
+var atomicHelpers = require( '../modules/util/atomic-helpers' );
 
 /**
  * ClearableInput
@@ -19,8 +16,7 @@ var atomicCheckers = require( '../modules/util/atomic-checkers' );
 function ClearableInput( element ) {
   var BASE_CLASS = 'input-contains-label';
 
-  var _dom =
-    atomicCheckers.validateDomElement( element, BASE_CLASS, 'ClearableInput' );
+  var _dom = atomicHelpers.checkDom( element, BASE_CLASS, 'ClearableInput' );
   var _inputDom = _dom.querySelector( 'input' );
   var _clearBtnDom = _dom.querySelector( '.' + BASE_CLASS + '_after__clear' );