+++ cfgov/unprocessed/js/molecules/GlobalSearch.js
@@ -1,10 +1,7 @@
 'use strict';
 
-// Required polyfills for IE9.
-if ( !Modernizr.classlist ) { require( '../modules/polyfill/class-list' ); } // eslint-disable-line no-undef, global-require, no-inline-comments, max-len
-
 // Required modules.
-var atomicCheckers = require( '../modules/util/atomic-checkers' );
+var atomicHelpers = require( '../modules/util/atomic-helpers' );
 var breakpointState = require( '../modules/util/breakpoint-state' );
 var ClearableInput = require( '../modules/ClearableInput' );
 var EventObserver = require( '../modules/util/EventObserver' );
@@ -24,12 +21,10 @@ function GlobalSearch( element ) { // eslint-disable-line max-statements, no-inl
 
   var BASE_CLASS = 'm-global-search';
 
-  var _dom =
-    atomicCheckers.validateDomElement( element, BASE_CLASS, 'GlobalSearch' );
+  var _dom = atomicHelpers.checkDom( element, BASE_CLASS, 'GlobalSearch' );
   var _triggerSel = '.' + BASE_CLASS + '_trigger';
   var _triggerDom = _dom.querySelector( _triggerSel );
-  var _flyoutMenu =
-    new FlyoutMenu( _dom, _triggerSel, '.' + BASE_CLASS + '_content' ).init();
+  var _flyoutMenu = new FlyoutMenu( _dom ).init();
   var _contentDom = _dom.querySelector( '.' + BASE_CLASS + '_content' );
   var _searchInputDom;
   var _searchBtnDom;
@@ -59,9 +54,8 @@ function GlobalSearch( element ) { // eslint-disable-line max-statements, no-inl
     var clearableInput = new ClearableInput( inputContainsLabel );
     clearableInput.init();
 
-    _flyoutMenu.addEventListener( 'toggle',
-                                  _handleToggle.bind( this ) );
-    _flyoutMenu.addEventListener( 'expandBegin', _handleExpandBegin );
+    var handleExpandBeginBinded = _handleExpandBegin.bind( this );
+    _flyoutMenu.addEventListener( 'expandBegin', handleExpandBeginBinded );
     _flyoutMenu.addEventListener( 'collapseBegin', _handleCollapseBegin );
     _flyoutMenu.addEventListener( 'collapseEnd', _handleCollapseEnd );
 
@@ -136,18 +130,11 @@ function GlobalSearch( element ) { // eslint-disable-line max-statements, no-inl
   }
 
   /**
-   * Event handler for when the search input flyout is toggled,
-   * which opens/closes the search input.
-   */
-  function _handleToggle() {
-    this.dispatchEvent( 'toggle', { target: this } );
-  }
-
-  /**
    * Event handler for when FlyoutMenu expand transition begins.
    * Use this to perform post-expandBegin actions.
    */
   function _handleExpandBegin() {
+    this.dispatchEvent( 'expandBegin', { target: this } );
     if ( _isInDesktop() ) { _triggerDom.classList.add( 'u-hidden' ); }
     _contentDom.classList.remove( 'u-invisible' );
     _searchInputDom.select();