+++ cfgov/unprocessed/js/organisms/Header.js
@@ -1,7 +1,7 @@
 'use strict';
 
 // Required modules.
-var atomicCheckers = require( '../modules/util/atomic-checkers' );
+var atomicHelpers = require( '../modules/util/atomic-helpers' );
 var GlobalSearch = require( '../molecules/GlobalSearch.js' );
 var MegaMenu = require( '../organisms/MegaMenu.js' );
 
@@ -19,21 +19,25 @@ function Header( element ) {
 
   var BASE_CLASS = 'o-header';
 
-  var _dom =
-    atomicCheckers.validateDomElement( element, BASE_CLASS, 'Header' );
+  var _dom = atomicHelpers.checkDom( element, BASE_CLASS, 'Header' );
   var _globalSearch;
   var _megaMenu;
+  var _overlay;
 
   /**
+   * @param {HTMLNode} overlay
+   *   Overlay to show/hide when mobile mega menu is shown.
    * @returns {Object} The Header instance.
    */
-  function init() {
+  function init( overlay ) {
+    _overlay = overlay;
     _globalSearch = new GlobalSearch( _dom );
-    _globalSearch.addEventListener( 'toggle', _searchClicked );
+    _globalSearch.addEventListener( 'expandBegin', _searchExpandBegin );
     _globalSearch.init();
 
     _megaMenu = new MegaMenu( _dom );
-    _megaMenu.addEventListener( 'triggerClick', _megaMenuClicked );
+    _megaMenu.addEventListener( 'rootExpandBegin', _megaMenuExpandBegin );
+    _megaMenu.addEventListener( 'rootCollapseEnd', _megaMenuCollapseEnd );
     _megaMenu.init();
 
     return this;
@@ -42,15 +46,26 @@ function Header( element ) {
   /**
    * Handler for opening the search.
    */
-  function _searchClicked() {
+  function _searchExpandBegin() {
     _megaMenu.collapse();
   }
 
+
   /**
-   * Handler for opening the mega menu.
+   * Handler for when the mega menu begins expansion.
+   * Collapse the global search.
    */
-  function _megaMenuClicked() {
+  function _megaMenuExpandBegin() {
     _globalSearch.collapse();
+    _overlay.classList.remove( 'u-hidden' );
+  }
+
+  /**
+   * Handler for when the mega menu ends collapsing.
+   * Show an overlay.
+   */
+  function _megaMenuCollapseEnd() {
+    _overlay.classList.add( 'u-hidden' );
   }
 
   this.init = init;