+++ cfgov/unprocessed/js/organisms/MegaMenu.js
@@ -1,7 +1,7 @@
 'use strict';
 
 // Required modules.
-var atomicCheckers = require( '../modules/util/atomic-checkers' );
+var atomicHelpers = require( '../modules/util/atomic-helpers' );
 var breakpointState = require( '../modules/util/breakpoint-state' );
 var EventObserver = require( '../modules/util/EventObserver' );
 var FlyoutMenu = require( '../modules/FlyoutMenu' );
@@ -17,14 +17,10 @@ var FlyoutMenu = require( '../modules/FlyoutMenu' );
  * @returns {Object} An MegaMenu instance.
  */
 function MegaMenu( element ) {
-
   var BASE_CLASS = 'o-mega-menu';
 
-  var _dom =
-    atomicCheckers.validateDomElement( element, BASE_CLASS, 'MegaMenu' );
-  var _flyoutMenu = new FlyoutMenu( _dom,
-                                    '.' + BASE_CLASS + '_trigger',
-                                    '.' + BASE_CLASS + '_content' ).init();
+  var _dom = atomicHelpers.checkDom( element, BASE_CLASS, 'MegaMenu' );
+  var _flyoutMenu = new FlyoutMenu( _dom ).init();
   var _activeMenu = _flyoutMenu;
   var _activeMenuDom = _flyoutMenu.getDom().content;
   var _menuItems = _dom.querySelectorAll( '.' + BASE_CLASS + '_content-item' );
@@ -42,13 +38,13 @@ function MegaMenu( element ) {
     var initEventsBinded = _initEvents.bind( this );
     var menuItem;
     var submenu;
+    var childSel = '.' + BASE_CLASS + '_content-link__has-children';
     for ( var i = 1, len = _menuItems.length; i < len; i++ ) {
       menuItem = _menuItems[i];
-      if ( menuItem.querySelector( '.u-link__disabled' ) === null ) {
+      if ( menuItem.querySelector( '.u-link__disabled' ) === null &&
+           menuItem.querySelector( childSel ) !== null ) {
         submenu =
-          new FlyoutMenu( menuItem, '.' + BASE_CLASS + '_content-link',
-                          '.' + BASE_CLASS + '_subcontent',
-                          '.' + BASE_CLASS + '_subcontent-btn__back' ).init();
+          new FlyoutMenu( menuItem ).init();
         _subMenus[menuItem] = submenu;
         initEventsBinded( submenu );
       }
@@ -64,9 +60,9 @@ function MegaMenu( element ) {
    * @param {FlyoutMenu} menu - The FlyoutMenu to add event listeners to.
    */
   function _initEvents( menu ) {
-    menu.addEventListener( 'toggle', _handleToggle.bind( this ) );
-    menu.addEventListener( 'expandBegin', _handleExpandBegin );
+    menu.addEventListener( 'expandBegin', _handleExpandBegin.bind( this ) );
     menu.addEventListener( 'collapseBegin', _handleCollapseBegin );
+    menu.addEventListener( 'collapseEnd', _handleCollapseEnd.bind( this ) );
   }
 
   /**
@@ -144,6 +140,7 @@ function MegaMenu( element ) {
    * @returns {Object} A MegaMenu instance.
    */
   function collapse() {
+    _flyoutMenu.collapse();
     _activeMenu.collapse();
 
     return this;
@@ -152,24 +149,28 @@ function MegaMenu( element ) {
   /**
    * Event handler for when the search input flyout is toggled,
    * which opens/closes the search input.
-   * @param {Event} event - Event dispatched by the FlyoutMenu instance.
+   * @param {FlyoutMenu} target - menu that is expanding or collapsing.
    */
-  function _handleToggle( event ) {
-    var target = event.target;
+  function _handleToggle( target ) {
     if ( target === _flyoutMenu &&
          _activeMenu !== target ) {
       _activeMenu.collapse();
     }
     _activeMenu = target;
     _activeMenuDom = _activeMenu.getDom().content;
-    this.dispatchEvent( 'toggle', { target: this } );
   }
 
   /**
    * Event handler for when FlyoutMenu expand transition begins.
    * Use this to perform post-expandBegin actions.
+   * @param {Event} event - A FlyoutMenu event.
    */
-  function _handleExpandBegin() {
+  function _handleExpandBegin( event ) {
+    var target = event.target;
+    _handleToggle( target );
+    if ( target === _flyoutMenu ) {
+      this.dispatchEvent( 'rootExpandBegin', { target: this } );
+    }
     // If on a submenu, focus the back button, otherwise focus the first link.
     var firstMenuLink;
     if ( _activeMenu === _flyoutMenu ) {
@@ -185,8 +186,22 @@ function MegaMenu( element ) {
   /**
    * Event handler for when FlyoutMenu collapse transition begins.
    * Use this to perform post-collapseBegin actions.
+   * @param {Event} event - A FlyoutMenu event.
    */
-  function _handleCollapseBegin() {
+  function _handleCollapseBegin( event ) {
+    _handleToggle( event.target );
+    document.body.removeEventListener( 'mousedown', _handleBodyClick );
+  }
+
+  /**
+   * Event handler for when FlyoutMenu collapse transition ends.
+   * Use this to perform post-collapseEnd actions.
+   * @param {Event} event - A FlyoutMenu event.
+   */
+  function _handleCollapseEnd( event ) {
+    if ( event.target === _flyoutMenu ) {
+      this.dispatchEvent( 'rootCollapseEnd', { target: this } );
+    }
     document.body.removeEventListener( 'mousedown', _handleBodyClick );
   }