+++ cfgov/unprocessed/js/routes/on-demand/filterable-list-controls.js
@@ -1,0 +1,9 @@
+/* ==========================================================================
+   Scripts for Filterable List Controls
+   ========================================================================== */
+
+'use strict';
+
+// List of organisms used.
+var FilterableListControls = require( '../../organisms/FilterableListControls' );
+var filterableListControls = new FilterableListControls( document.body ).init();